pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "First steps with LangChain"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langchain-openai>=1.1.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]
